Configurable + Multiple application source

+++ phoenicis-tools/src/main/java/com/playonlinux/tools/http/Downloader.java
@@ -90,7 +90,7 @@ public class Downloader {
                 totalDataRead += i;
                 bufferedOutputStream.write(data, 0, i);
 
-                percentage = totalDataRead * 100 / fileSize;
+                percentage = ((float) totalDataRead * 100) / ((float) fileSize);
 
                 changeState(
                         ProgressState.PROGRESSING, percentage,
